tests/wpt/web-platform-tests/css/css-color/parsing/color-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Level 3: parsing color with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-color-3/#foreground">
<meta name="assert" content="color supports the '<color>' grammar defined in CSS Color Level 3.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("color", "currentcolor");
test_valid_value("color", "transparent");
test_valid_value("color", "red");
test_valid_value("color", "magenta");
test_valid_value("color", "#234", "rgb(34, 51, 68)");
test_valid_value("color", "#FEDCBA", "rgb(254, 220, 186)");
test_valid_value("color", "rgb(2, 3, 4)");
test_valid_value("color", "rgb(100%, 0%, 0%)", "rgb(255, 0, 0)");
test_valid_value("color", "rgba(2, 3, 4, 0.5)");
test_valid_value("color", "rgba(2, 3, 4, 50%)", "rgba(2, 3, 4, 0.5)");
test_valid_value("color", "hsl(120, 100%, 50%)", "rgb(0, 255, 0)");
test_valid_value("color", "hsla(120, 100%, 50%, 0.25)", "rgba(0, 255, 0, 0.25)");
test_valid_value("color", "rgb(-2, 3, 4)", "rgb(0, 3, 4)");
test_valid_value("color", "rgb(100, 200, 300)", "rgb(100, 200, 255)");
test_valid_value("color", "rgb(20, 10, 0, -10)", "rgba(20, 10, 0, 0)");
test_valid_value("color", "rgb(100%, 200%, 300%)", "rgb(255, 255, 255)");
</script>
</body>
</html>
